+++ test/libsolidity/smtCheckerTests/types/tuple_extra_parens_7.sol
@@ -4,7 +4,7 @@ contract C {
 		return (2, 3);
 	}
 	function f() public {
-		(address(1).call(""));
+		(address(uint160(1)).call(""));
 		(uint x, uint y) = ((g()));
 		assert(x == 2);
 		assert(y == 3);
